feat: Update API constants and dependencies
Update the API constants to use the new secret API key for the base URL. Also, update the Firebase dependencies to their latest versions: firebase_auth (5.0.0), firebase_core (3.0.0), cloud_firestore (5.0.0), and firebase_storage (12.0.0). Additionally, update the minSdkVersion to 23 and targetSdkVersion to 34 in the Android build.gradle file. Finally, add the loginBehavior parameter with the value LoginBehavior.webOnly to the Facebook Sign-In functionality in the authentication_repository.dart file.